chore: move scratch_blocks_utils.js into src
