+++ src/index.js
@@ -22,7 +22,7 @@ import "../blocks_vertical/operators.js";
 import "../blocks_vertical/procedures.js";
 import "../blocks_vertical/sensing.js";
 import "../blocks_vertical/sound.js";
-import * as scratchBlocksUtils from "../core/scratch_blocks_utils.js";
+import * as scratchBlocksUtils from "./scratch_blocks_utils.js";
 import * as ScratchVariables from "./variables.js";
 import "./css.js";
 import "./fields/field_vertical_separator.js";
